Ludus-Bot/requirements.txt
discord.py>=2.3.0
yt-dlp>=2024.12.23
PyNaCl>=1.5.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
Pillow>=10.0.0
aiofiles>=23.2.1
